% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_tutorial}
\alias{load_tutorial}
\title{Loads an R tutorial included in CKteachR}
\usage{
load_tutorial(tutorial_name, tutorial_shiny_args = NULL)
}
\arguments{
\item{tutorial_name}{The tutorial name. See \code{\link[=list_tutorials]{list_tutorials()}} for a list of
available tutorials.}

\item{tutorial_shiny_args}{This pass directly to the \code{\link[learnr:run_tutorial]{learnr::run_tutorial()}}
function from the learnr package.}
}
\description{
This function is a wrapper for the \code{\link[learnr:run_tutorial]{learnr::run_tutorial()}} function from the
learnr package. See the documentation and vignettes of that package.
}
