% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setup_progress_monitoring}
\alias{setup_progress_monitoring}
\title{Sets up the progress monitoring environment}
\usage{
setup_progress_monitoring(
  working_dir,
  tutorial_name,
  students_gs4_ident,
  mark_gs4_ident,
  mark_gs4_email
)
}
\arguments{
\item{working_dir}{The path to the working directory.}

\item{tutorial_name}{The tutorial_name.}

\item{students_gs4_ident}{The Googlesheet ID with student credentials.}

\item{mark_gs4_ident}{The Googlesheet ID that will record students' progress.}

\item{mark_gs4_email}{The email linked to the Google account}
}
\value{
A list with the necessary functions for \code{submitr::shiny_logic()}.
}
\description{
This function sets up the progress monitoring environment. If a secret key
exists in the working dir but not an decrypted token, a decrypted token is
created. If a decrypted token exists
}
